+++ src/components/common/button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import Button from './Button';
-import Margin from '../layout/Margin';
+import Margin from '../../layout/Margin';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
