+++ src/components/common/button/ButtonSet.tsx
@@ -1,6 +1,5 @@
 import { ReactNode } from 'react';
 import styled from 'styled-components';
-import Button from '../common/Button';
 
 interface ButtonSetProps {
   /**
